srs/requirements.txt
Pillow
colorlog
cryptography
ebooklib
flask
gunicorn
matplotlib
psycopg2-binary
pydub
pymupdf
python-docx
python-dotenv
python-pptx
requests
schedule
